requirements.txt
APScheduler==3.10.4
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
eventlet==0.33.3
Flask==2.3.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-engineio==4.4.1
python-socketio==5.7.2
pytz==2024.2
simple-websocket==1.1.0
six==1.16.0
tzdata==2024.2
tzlocal==5.2
uvicorn==0.22.0
Werkzeug>=2.2.0  # Update to a compatible version
wsproto==1.2.0
gunicorn==20.1.0

